requirements.txt
# Framework Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Base de données
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Validation et sérialisation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentification
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# HTTP clients
httpx==0.26.0
aiohttp==3.9.1

# TTS
piper-tts==1.2.0

# GPIO (Raspberry Pi)
RPi.GPIO==0.7.1; platform_machine == "armv7l" or platform_machine == "aarch64"

# Utilitaires
python-dateutil==2.8.2
pytz==2023.3

# Tests
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
